Add > >= < <= for numbers per recommendation from bogon64.
While implementing noticed that I leak false vs nil in lots of places.
The LUA code uses false in lots of places, but if it leaks into pico8lisp we crash.
Because nil in LUA is falsey I was able to fix the issue by coercing false to nil.
Almost all == true and == false have been removed.